requirements.txt
# requirements

Django==4.1
django-cors-headers==3.13.0
PyMySQL==1.0.2
mysqlclient==2.1.1
PyJWT==2.4.0
